Results 1. js에서 쿠키 핸들링 못하도록 secure 속성 설정 => 기존 화면단에서 accessToken 체크하는 방식을 서버 인터셉터에서 체크하도록 변경 2. 이메일 전송시 콤마로 이어붙일시 한 번에 여러 개의 주소로 발송 가능하도록 개선 3. 로컬 이외의 모든 cors 요청을 허용하도록 변경 4. Vuex 라이브러리 추가 - 화면단에서 로그인 상태 관리하는 전역변수 설정 TODO 1. yes24 크롤링해와서 검색 만으로 책 추가할 수 있도록 구현(이미지 포함) 2. 새로운 문장 추가시 해당 키를 기존 로컬스토리지 sentenceSortKey 값에 업데이트 해주도록 구현. 마찬가지로 문장 삭제시에도 키 삭제 후 업데이트 3. 단일 문장 추가 기능 구현 4. 로그아웃 구현 5. 일반 로그인 방식 및 회원가입 기능 구현 6. 화면단 전체적인 삭제 기능 구현 7. 책 상세페이지에서 문장 즐겨찾기 기능 구현(favorite_flag) 이후 즐겨찾기 한 기능 모아볼 수 있는 기능도 추가 8. 사용자의 책 데이터를 검색하는 기능 추가(태그 검색도) 9. (후순위) 메인화면 로드시마다 랜덤 문장 하나씩 가져와 노출시키는 기능. css 애니메이션 사용해서 구현해보기

+++ bookMemory/src/main/java/com/yh/bookMemory/dto/UserDTO.java
@@ -1,7 +1,6 @@
 package com.yh.bookMemory.dto;
 
 import com.yh.bookMemory.entity.Users;
-import jakarta.persistence.Column;
 import lombok.Builder;
 import lombok.Data;
 
@@ -21,7 +20,7 @@ public class UserDTO {
     private Double sortKey = 0.0; // 랜덤 문장 선정할 때 order by 할 난수
 
     @Builder.Default
-    private Long sendCnt = 0L;    // 랜덤 문장 받은 개수
+    private Integer sendCnt = 0;    // 랜덤 문장 받은 개수
 
     public Users toEntity() {
         return Users.builder()
